requirements.txt
# --------- pytorch --------- #
torch>=1.11.0
torchvision>=0.11.0
pytorch-lightning>=1.5.10
torchmetrics>=0.7.0

# --------- diff eq --------- #
torchdiffeq==0.2.3
scipy==1.8.0
scikit-learn==1.0.2

# --------- plotting --------- #
matplotlib==3.5.0

# --------- hydra --------- #
hydra-core>=1.1.0
hydra-colorlog>=1.1.0
hydra-optuna-sweeper>=1.1.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml
tensorboard

# --------- linters --------- #
pre-commit      # hooks for applying linters on commit
black           # code formatting
isort           # import sorting
flake8          # code analysis
nbstripout      # remove output from jupyter notebooks

# --------- others --------- #
python-dotenv   # loading env variables from .env file
rich            # beautiful text formatting in terminal
pytest          # tests
sh              # for running bash commands in some tests
pudb            # debugger

# --------- datasets --------- #
xarray==2022.3.0
dask==2022.4.1
netCDF4==1.5.8
pydap==3.2.2 # for loading data from OPeNDAP

# --------- models --------- #
sympy # required for pdenet
